支持Android平台构建.so而不是默认的.aar；win64平台仅生成Realease以加快构建

+++ src/bq_log/api/unity_log_plugin.cpp
@@ -1,12 +1,14 @@
 #include "UnityPluginAPI/IUnityLog.h"
 #include <bq_log/bq_log.h>
+#include <stdint.h>
+
 static IUnityLog* s_Logger;
 
 static void InitBqLog();
 static void UnInitBqLog();
 static bq::log GetBqLog();
 
-static void OnBqLog(unsigned long long logId, int categoryIndex, bq::log_level logLevel, 
+static void OnBqLog(uint64_t logId, int categoryIndex, bq::log_level logLevel, 
 	const char* content, int length)
 {
 	(void)length;
